+++ zone/lua_mob.h
@@ -310,8 +310,8 @@ public:
 		uint32 unk020, bool perm_effect, Lua_Client c);
 	void TempName();
 	void TempName(const char *newname);
-	void SetGlobal(const char *varname, const char *newvalue, int options, const char *duration);
-	void SetGlobal(const char *varname, const char *newvalue, int options, const char *duration, Lua_Mob other);
+	void SetGlobal(const char *varname, const char *newvalue, uint8 options, const char *duration);
+	void SetGlobal(const char *varname, const char *newvalue, uint8 options, const char *duration, Lua_Mob other);
 	void TarGlobal(const char *varname, const char *value, const char *duration, int npc_id, int char_id, int zone_id);
 	void DelGlobal(const char *varname);
 	void SetSlotTint(int material_slot, int red_tint, int green_tint, int blue_tint);
